refactor: split ContextApi (get/init → support, snapshot → profiler)
