+++ ddprof-lib/src/main/cpp/javaApi.cpp
@@ -19,7 +19,7 @@
 
 #include "arch.h"
 #include "context.h"
-#include "context_api.h"
+#include "support/context_api.h"
 #include "guards.h"
 #include "counters.h"
 #include "common.h"
